Move methods back to OidcAuthenticationService
- Implement suggestions
