+++ alpine-server/src/main/java/alpine/server/auth/OidcAuthenticationService.java
@@ -21,10 +21,14 @@ package alpine.server.auth;
 
 import alpine.Config;
 import alpine.common.logging.Logger;
+import alpine.model.OidcUser;
+import alpine.persistence.AlpineQueryManager;
 import alpine.server.util.OidcUtil;
 
 import jakarta.annotation.Nonnull;
 import java.security.Principal;
+import java.util.List;
+import java.util.Objects;
 import java.util.ServiceLoader;
 
 /**
@@ -63,13 +67,8 @@ public class OidcAuthenticationService implements AuthenticationService {
     /**
      * Constructor for unit tests
      */
-    OidcAuthenticationService(final Config config,
-            final OidcConfiguration oidcConfiguration,
-            final String idToken,
-            final String accessToken) {
-        this(config, oidcConfiguration,
-                new OidcIdTokenAuthenticator(oidcConfiguration, config.getProperty(Config.AlpineKey.OIDC_CLIENT_ID)),
-                new OidcUserInfoAuthenticator(oidcConfiguration), idToken, accessToken);
+    OidcAuthenticationService(final Config config, final OidcConfiguration oidcConfiguration, final String idToken, final String accessToken) {
+        this(config, oidcConfiguration, new OidcIdTokenAuthenticator(oidcConfiguration, config.getProperty(Config.AlpineKey.OIDC_CLIENT_ID)), new OidcUserInfoAuthenticator(oidcConfiguration), idToken, accessToken);
     }
 
     /**
@@ -78,11 +77,11 @@ public class OidcAuthenticationService implements AuthenticationService {
      * @since 1.10.0
      */
     OidcAuthenticationService(final Config config,
-            final OidcConfiguration oidcConfiguration,
-            final OidcIdTokenAuthenticator idTokenAuthenticator,
-            final OidcUserInfoAuthenticator userInfoAuthenticator,
-            final String idToken,
-            final String accessToken) {
+                              final OidcConfiguration oidcConfiguration,
+                              final OidcIdTokenAuthenticator idTokenAuthenticator,
+                              final OidcUserInfoAuthenticator userInfoAuthenticator,
+                              final String idToken,
+                              final String accessToken) {
         this.config = config;
         this.oidcConfiguration = oidcConfiguration;
         this.idTokenAuthenticator = idTokenAuthenticator;
@@ -101,7 +100,8 @@ public class OidcAuthenticationService implements AuthenticationService {
 
     @Override
     public boolean isSpecified() {
-        return OidcUtil.isOidcAvailable(config, oidcConfiguration) && (accessToken != null || idToken != null);
+        return OidcUtil.isOidcAvailable(config, oidcConfiguration)
+                && (accessToken != null || idToken != null);
     }
 
     /**
@@ -130,7 +130,6 @@ public class OidcAuthenticationService implements AuthenticationService {
 
         final boolean teamSyncEnabled = config.getPropertyAsBoolean(Config.AlpineKey.OIDC_TEAM_SYNCHRONIZATION);
         final String teamsClaimName = config.getProperty(Config.AlpineKey.OIDC_TEAMS_CLAIM);
-
         if (teamSyncEnabled && teamsClaimName == null) {
             LOGGER.error("Team synchronization is enabled, but no teams claim has been configured");
             throw new AlpineAuthenticationException(AlpineAuthenticationException.CauseType.OTHER);
@@ -147,7 +146,7 @@ public class OidcAuthenticationService implements AuthenticationService {
 
             if (customizer.isProfileComplete(idTokenProfile, teamSyncEnabled)) {
                 LOGGER.debug("ID token profile is complete, proceeding to authenticate");
-                return customizer.onAuthenticationSuccess(idTokenProfile);
+                return authenticateInternal(idTokenProfile);
             }
         }
 
@@ -158,7 +157,7 @@ public class OidcAuthenticationService implements AuthenticationService {
 
             if (customizer.isProfileComplete(userInfoProfile, teamSyncEnabled)) {
                 LOGGER.debug("UserInfo profile is complete, proceeding to authenticate");
-                return customizer.onAuthenticationSuccess(userInfoProfile);
+                return authenticateInternal(userInfoProfile);
             }
         }
 
@@ -169,14 +168,73 @@ public class OidcAuthenticationService implements AuthenticationService {
 
             if (customizer.isProfileComplete(mergedProfile, teamSyncEnabled)) {
                 LOGGER.debug("Merged profile is complete, proceeding to authenticate");
-                return customizer.onAuthenticationSuccess(mergedProfile);
+                return authenticateInternal(mergedProfile);
             }
         }
 
-        LOGGER.error("Unable to assemble complete profile (ID token: %s, UserInfo: %s, Merged: %s)"
-                .formatted(idTokenProfile, userInfoProfile, mergedProfile));
-
+        LOGGER.error("Unable to assemble complete profile (ID token: " + idTokenProfile +
+                ", UserInfo: " + userInfoProfile + ", Merged: " + mergedProfile + ")");
         throw new AlpineAuthenticationException(AlpineAuthenticationException.CauseType.OTHER);
+    }
+
+    private OidcUser authenticateInternal(final OidcProfile profile) throws AlpineAuthenticationException {
+        try (final var qm = new AlpineQueryManager()) {
+            OidcUser user = qm.getOidcUser(profile.getUsername());
+            if (user != null) {
+                LOGGER.debug("Attempting to authenticate user: " + user.getUsername());
+                if (user.getSubjectIdentifier() == null) {
+                    LOGGER.debug("Assigning subject identifier " + profile.getSubject() + " to user " + user.getUsername());
+                    user.setSubjectIdentifier(profile.getSubject());
+                    user.setEmail(profile.getEmail());
+                    customizer.onAuthenticationSuccess(profile, idToken, accessToken);
+                    return qm.updateOidcUser(user);
+                } else if (!user.getSubjectIdentifier().equals(profile.getSubject())) {
+                    LOGGER.error("Refusing to authenticate user " + user.getUsername() + ": subject identifier has changed (" +
+                            user.getSubjectIdentifier() + " to " + profile.getSubject() + ")");
+                    throw new AlpineAuthenticationException(AlpineAuthenticationException.CauseType.INVALID_CREDENTIALS);
+                }
+                if (!Objects.equals(user.getEmail(), profile.getEmail())) {
+                    LOGGER.debug("Updating email of user " + user.getUsername() + ": " + user.getEmail() + " -> " + profile.getEmail());
+                    user.setEmail(profile.getEmail());
+                    user = qm.updateOidcUser(user);
+                }
+                if (config.getPropertyAsBoolean(Config.AlpineKey.OIDC_TEAM_SYNCHRONIZATION)) {
+                    customizer.onAuthenticationSuccess(profile, idToken, accessToken);
+                    return qm.synchronizeTeamMembership(user, profile.getGroups());
+                }
+                customizer.onAuthenticationSuccess(profile, idToken, accessToken);
+                return user;
+            } else if (config.getPropertyAsBoolean(Config.AlpineKey.OIDC_USER_PROVISIONING)) {
+                LOGGER.debug("The user (" + profile.getUsername() + ") authenticated successfully but the account has not been provisioned");
+                customizer.onAuthenticationSuccess(profile, idToken, accessToken);
+                return autoProvision(qm, profile);
+            } else {
+                LOGGER.debug("The user (" + profile.getUsername() + ") is unmapped and user provisioning is not enabled");
+                throw new AlpineAuthenticationException(AlpineAuthenticationException.CauseType.UNMAPPED_ACCOUNT);
+            }
+        }
+    }
+
+    private OidcUser autoProvision(final AlpineQueryManager qm, final OidcProfile profile) {
+        var user = new OidcUser();
+        user.setUsername(profile.getUsername());
+        user.setSubjectIdentifier(profile.getSubject());
+        user.setEmail(profile.getEmail());
+        user = qm.persist(user);
+
+        if (config.getPropertyAsBoolean(Config.AlpineKey.OIDC_TEAM_SYNCHRONIZATION)) {
+            LOGGER.debug("Synchronizing teams for user " + user.getUsername());
+            customizer.onAuthenticationSuccess(profile, idToken, accessToken);
+            return qm.synchronizeTeamMembership(user, profile.getGroups());
+        }
+
+        final List<String> defaultTeams = config.getPropertyAsList(Config.AlpineKey.OIDC_TEAMS_DEFAULT);
+        if (!defaultTeams.isEmpty()) {
+            LOGGER.debug("Assigning default teams %s to user %s".formatted(defaultTeams, user.getUsername()));
+            return qm.addUserToTeams(user, defaultTeams);
+        }
+
+        return user;
     }
 
 }
